change(animations): Move animations, show them in styleguide
