+++ src/animations/index.js
@@ -1,1 +1,3 @@
+export fullCircleRotate from './fullCircleRotate';
 export pulseInfinite from './pulseInfinite';
+export slideIn from './slideIn';
