+++ src/animations/pulseInfinite.js
@@ -1,7 +1,7 @@
 import { keyframes } from 'styled-components';
 
 /** @component */
-const pulseInfinite = keyframes`
+const pulseInfinite = () => keyframes`
   40% { transform: scale(1); }
   60% { transform: scale(1.2); }
   80% { transform: scale(1); }
